refactor(BSQ) Optimize writing and free heap
Writing is now done using that maps's stat.st_size value (less the
lenght of the first line) and write() instead of fox_putstr to preven
having to read the characters, as the st_size and characters are checked
for validity beforehand.

Also, the program frees stuff now. That's better. Valgrind is happy, so
am I.